Extract csv download to custom hook
